requirements.txt
timm==0.3.2
torchvision==0.12.0
numpy==1.19.5
nltk==3.7
scipy==1.7.3
torch==1.11.0
Pillow==9.4.0
scikit_learn==1.2.1
